fix(runtime-learning): move persistence helpers into runtime; keep noesis.learn as stable facade
Fix a Clean Architecture boundary violation where noesis/runtime/learning.py
imported filesystem persistence helpers from the public facade noesis/learn.py.

Changes:
- Move learning persistence + snapshot I/O helpers into noesis/runtime/learning.py:
  - build_learn_payload
  - persist_episode_learning (writes learn.jsonl)
  - load_policy_snapshot / update_policy_snapshot (policy snapshots under learn_home)
  - summarise_learn_kinds
- Convert noesis/learn.py into a thin facade that re-exports the same public symbols
  from runtime and preserves existing behavior (including emit alias).
- Add a minimal guard test to ensure runtime never imports noesis.learn again.

Public API:
- No user-facing import changes. `from noesis.learn import ...` continues to work.

Tests:
- pytest tests/public/test_forbidden_imports.py tests/public/test_public_imports.py
