+++ .gitignore
@@ -49,3 +49,4 @@ Thumbs.db
 ### Cursor
 rules.md
 docs/rules.md
+AGENTS.md
