+++ tests/public/test_forbidden_imports.py
@@ -24,6 +24,14 @@ def _iter_source_files() -> list[Path]:
     return files
 
 
+def _iter_runtime_files() -> list[Path]:
+    repo_root = Path(__file__).resolve().parents[2]
+    runtime_root = repo_root / "noesis" / "runtime"
+    if not runtime_root.exists():
+        return []
+    return [path for path in runtime_root.rglob("*.py") if path.is_file()]
+
+
 def _should_check_line(path: Path, line: str, state: dict) -> bool:
     if path.suffix not in {".md", ".mdx"}:
         return True
@@ -56,3 +64,13 @@ def test_docs_and_examples_do_not_import_private_modules() -> None:
             if any(prefix in stripped for prefix in FORBIDDEN_PREFIXES):
                 offenders.append(f"{path}: {stripped}")
     assert not offenders, "Forbidden imports detected:\n" + "\n".join(offenders)
+
+
+def test_runtime_does_not_import_noesis_learn() -> None:
+    offenders: list[str] = []
+    for path in _iter_runtime_files():
+        for line in path.read_text(encoding="utf-8").splitlines():
+            stripped = line.strip()
+            if stripped.startswith(("import ", "from ")) and "noesis.learn" in stripped:
+                offenders.append(f"{path}: {stripped}")
+    assert not offenders, "Runtime imports noesis.learn:\n" + "\n".join(offenders)
